+++ feature/main/src/main/java/com/umcspot/spot/main/MainNavHost.kt
@@ -23,7 +23,7 @@ import com.umcspot.spot.feature.board.post.posting.navigation.postingGraph
 import com.umcspot.spot.home.navigation.homeGraph
 import com.umcspot.spot.jjim.navigation.jjimGraph
 import com.umcspot.spot.model.QuickMenuType
-import com.umcspot.spot.mypage.navigation.mypageGraph
+import com.umcspot.spot.mypage.main.navigation.mypageGraph
 import com.umcspot.spot.signup.navigation.signupGraph
 import com.umcspot.spot.study.detail.navigation.navigateToStudyDetail
 import com.umcspot.spot.study.detail.navigation.studyDetailGraph
