+++ feature/main/src/main/java/com/umcspot/spot/main/MainNavTab.kt
@@ -16,7 +16,7 @@ import com.umcspot.spot.main.R.string.ic_category_desc
 import com.umcspot.spot.main.R.string.ic_mystudy_desc
 import com.umcspot.spot.main.R.string.ic_jjim_desc
 import com.umcspot.spot.main.R.string.ic_mypage_desc
-import com.umcspot.spot.mypage.navigation.MyPage
+import com.umcspot.spot.mypage.main.navigation.MyPage
 import com.umcspot.spot.study.my.navigation.MyStudy
 import com.umcspot.spot.navigation.MainTabRoute
 import com.umcspot.spot.navigation.Route
