+++ feature/mypage/src/main/java/com/umcspot/spot/mypage/main/MyPageScreen.kt
@@ -1,4 +1,4 @@
-package com.umcspot.spot.mypage
+package com.umcspot.spot.mypage.main
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -49,7 +49,6 @@ import com.umcspot.spot.designsystem.theme.G400
 import com.umcspot.spot.designsystem.theme.SpotTheme
 import com.umcspot.spot.model.ImageRef
 import com.umcspot.spot.model.SocialLoginType
-import com.umcspot.spot.model.StudyTheme
 import com.umcspot.spot.ui.extension.screenHeightDp
 import com.umcspot.spot.ui.extension.screenWidthDp
 import com.umcspot.spot.ui.state.UiState
