+++ feature/mypage/src/main/java/com/umcspot/spot/mypage/main/navigation/MypageNavigation.kt
@@ -1,11 +1,11 @@
-package com.umcspot.spot.mypage.navigation
+package com.umcspot.spot.mypage.main.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
-import com.umcspot.spot.mypage.MyPageScreen
+import com.umcspot.spot.mypage.main.MyPageScreen
 import com.umcspot.spot.navigation.MainTabRoute
 import kotlinx.serialization.Serializable
